@@ -25,57 +25,57 @@
     }
   ],
   "dependencies": {
-    "@angular/animations": "4.3.4",
-    "@angular/cdk": "2.0.0-beta.8",
-    "@angular/cli": "1.3.0",
-    "@angular/common": "4.3.4",
-    "@angular/compiler": "4.3.4",
-    "@angular/compiler-cli": "4.3.4",
-    "@angular/core": "4.3.4",
-    "@angular/forms": "4.3.4",
-    "@angular/http": "4.3.4",
-    "@angular/material": "2.0.0-beta.8",
-    "@angular/platform-browser": "4.3.4",
-    "@angular/platform-browser-dynamic": "4.3.4",
-    "@angular/platform-server": "4.3.4",
-    "@angular/router": "4.3.4",
-    "@angular/upgrade": "4.3.4",
-    "angular-in-memory-web-api": "0.3.2",
-    "core-js": "2.5.0",
-    "mydatepicker": "2.0.27",
+    "@angular/animations": "5.1.1",
+    "@angular/cdk": "5.0.1",
+    "@angular/cli": "1.6.1",
+    "@angular/common": "5.1.1",
+    "@angular/compiler": "5.1.1",
+    "@angular/compiler-cli": "5.1.1",
+    "@angular/core": "5.1.1",
+    "@angular/forms": "5.1.1",
+    "@angular/http": "5.1.1",
+    "@angular/material": "5.0.1",
+    "@angular/platform-browser": "5.1.1",
+    "@angular/platform-browser-dynamic": "5.1.1",
+    "@angular/platform-server": "5.1.1",
+    "@angular/router": "5.1.1",
+    "@angular/upgrade": "5.1.1",
+    "angular-in-memory-web-api": "0.5.2",
+    "core-js": "2.5.3",
+    "mydatepicker": "2.6.1",
     "reflect-metadata": "0.1.10",
-    "rxjs": "5.4.3",
-    "systemjs": "0.20.17",
-    "zone.js": "0.8.16"
+    "rxjs": "5.5.5",
+    "systemjs": "0.20.19",
+    "zone.js": "0.8.18"
   },
   "devDependencies": {
-    "@ngtools/webpack": "1.6.0",
-    "angular2-autosize": "1.0.1",
+    "@ngtools/webpack": "1.9.1",
+    "angular2-autosize": "1.0.2",
     "angular2-template-loader": "0.6.2",
-    "awesome-typescript-loader": "3.2.2",
-    "compression-webpack-plugin": "1.0.0",
-    "concurrently": "3.5.0",
-    "copy-webpack-plugin": "4.0.1",
-    "css-loader": "0.28.4",
-    "extract-text-webpack-plugin": "3.0.0",
-    "file-loader": "0.11.2",
+    "awesome-typescript-loader": "3.4.1",
+    "compression-webpack-plugin": "1.1.2",
+    "concurrently": "3.5.1",
+    "copy-webpack-plugin": "4.3.0",
+    "css-loader": "0.28.7",
+    "extract-text-webpack-plugin": "3.0.2",
+    "file-loader": "1.1.6",
     "html-loader": "0.5.1",
     "html-webpack-plugin": "2.30.1",
-    "jasmine": "2.7.0",
-    "karma": "1.7.0",
-    "karma-jasmine": "1.1.0",
+    "jasmine": "2.8.0",
+    "karma": "1.7.1",
+    "karma-jasmine": "1.1.1",
     "karma-phantomjs-launcher": "1.0.4",
     "karma-sourcemap-loader": "0.3.7",
-    "karma-webpack": "2.0.4",
+    "karma-webpack": "2.0.9",
     "lite-server": "2.3.0",
     "open-browser-webpack-plugin": "0.0.5",
-    "postcss": "6.0.8",
+    "postcss": "6.0.14",
     "raw-loader": "0.5.1",
-    "style-loader": "0.18.2",
+    "style-loader": "0.19.1",
     "to-string-loader": "1.1.5",
     "typescript": "2.4.2",
-    "webpack": "3.5.3",
-    "webpack-dev-server": "2.7.1",
-    "webpack-merge": "4.1.0"
+    "webpack": "3.10.0",
+    "webpack-dev-server": "2.9.7",
+    "webpack-merge": "4.1.1"
   }
 }
