@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Response, Http } from '@angular/http';
+import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs/Observable';
 
 import { ManufacturerModel } from './manufacturer-model.model';
@@ -8,46 +8,46 @@ import { HandleErrorService } from '../../shared/utils/handle-error.service';
 
 @Injectable()
 export class ManufacturerModelService {
-    constructor(private http: Http, private handleErrorService: HandleErrorService) { }
+    constructor(private http: HttpClient, private handleErrorService: HandleErrorService) { }
 
     getManufacturerModels(): Promise<ManufacturerModel[]> {
-        return this.http.get(AppUtils.BACKEND_API_MANUF_MODEL_URL)
+        return this.http.get<ManufacturerModel[]>(AppUtils.BACKEND_API_MANUF_MODEL_URL)
             .toPromise()
-            .then(response => response.json() as ManufacturerModel[])
+            .then(response => response)
             .catch((error) => {
                 console.error('Error while getting manufModels', error);
                 return Promise.reject(error.message || error);
-        });
+            });
     }
 
-    delete(id: number): Promise<Response> {
-        return this.http.delete(AppUtils.BACKEND_API_MANUF_MODEL_URL + '/' + id)
+    delete(id: number): Promise<void> {
+        return this.http.delete<void>(AppUtils.BACKEND_API_MANUF_MODEL_URL + '/' + id)
             .toPromise()
             .catch((error) => {
                 console.error('Error delete manufModel', error);
                 return Promise.reject(error.message || error);
-        });
+            });
     }
 
-    getManufacturerModel (id: number): Promise<ManufacturerModel> {
-        return this.http.get(AppUtils.BACKEND_API_MANUF_MODEL_URL + '/' + id)
+    getManufacturerModel(id: number): Promise<ManufacturerModel> {
+        return this.http.get<ManufacturerModel>(AppUtils.BACKEND_API_MANUF_MODEL_URL + '/' + id)
             .toPromise()
-            .then(res => res.json() as ManufacturerModel)
+            .then(res => res)
             .catch((error) => {
                 console.error('Error while getting manufModel', error);
                 return Promise.reject(error.message || error);
-        });
+            });
     }
 
     create(manufModel: ManufacturerModel): Observable<ManufacturerModel> {
-        return this.http.post(AppUtils.BACKEND_API_MANUF_MODEL_URL, JSON.stringify(manufModel))
-            .map(response => response.json() as ManufacturerModel)
+        return this.http.post<ManufacturerModel>(AppUtils.BACKEND_API_MANUF_MODEL_URL, JSON.stringify(manufModel))
+            .map(response => response)
             .catch(this.handleErrorService.handleError);
     }
 
     update(id: number, manufModel: ManufacturerModel): Observable<ManufacturerModel> {
-        return this.http.put(AppUtils.BACKEND_API_MANUF_MODEL_URL + '/' + id, JSON.stringify(manufModel))
-            .map(response => response.json() as ManufacturerModel)
+        return this.http.put<ManufacturerModel>(AppUtils.BACKEND_API_MANUF_MODEL_URL + '/' + id, JSON.stringify(manufModel))
+            .map(response => response)
             .catch(this.handleErrorService.handleError);
     }
 }
