@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Response, Http } from '@angular/http';
+import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs/Observable';
 
 import * as AppUtils from '../../utils/app.utils';
@@ -9,20 +9,20 @@ import { SubjectExamination } from '../shared/subject-examination.model';
 
 @Injectable()
 export class ExaminationService {
-    constructor(private http: Http, private handleErrorService: HandleErrorService) { }
+    constructor(private http: HttpClient, private handleErrorService: HandleErrorService) { }
 
     getExaminations(): Promise<Examination[]> {
-        return this.http.get(AppUtils.BACKEND_API_EXAMINATION_URL)
+        return this.http.get<Examination[]>(AppUtils.BACKEND_API_EXAMINATION_URL)
             .toPromise()
-            .then(response => response.json() as Examination[])
+            .then(response => response)
             .catch((error) => {
                 console.error('Error while getting examinations', error);
                 return Promise.reject(error.message || error);
             });
     }
 
-    delete(id: number): Promise<Response> {
-        return this.http.delete(AppUtils.BACKEND_API_EXAMINATION_URL + '/' + id)
+    delete(id: number): Promise<void> {
+        return this.http.delete<void>(AppUtils.BACKEND_API_EXAMINATION_URL + '/' + id)
             .toPromise()
             .catch((error) => {
                 console.error('Error delete examination', error);
@@ -31,9 +31,9 @@ export class ExaminationService {
     }
 
     getExamination(id: number): Promise<Examination> {
-        return this.http.get(AppUtils.BACKEND_API_EXAMINATION_URL + '/' + id)
+        return this.http.get<Examination>(AppUtils.BACKEND_API_EXAMINATION_URL + '/' + id)
             .toPromise()
-            .then(res => res.json() as Examination)
+            .then(res => res)
             .catch((error) => {
                 console.error('Error while getting examination', error);
                 return Promise.reject(error.message || error);
@@ -41,21 +41,21 @@ export class ExaminationService {
     }
 
     create(examination: Examination): Observable<Examination> {
-        return this.http.post(AppUtils.BACKEND_API_EXAMINATION_URL, JSON.stringify(examination))
-            .map(this.handleErrorService.extractData)
+        return this.http.post<Examination>(AppUtils.BACKEND_API_EXAMINATION_URL, JSON.stringify(examination))
+            .map(response => response)
             .catch(this.handleErrorService.handleError);
     }
 
     update(id: number, examination: Examination): Observable<Examination> {
-        return this.http.put(AppUtils.BACKEND_API_EXAMINATION_URL + '/' + id, JSON.stringify(examination))
-            .map(response => response.json() as Examination)
+        return this.http.put<Examination>(AppUtils.BACKEND_API_EXAMINATION_URL + '/' + id, JSON.stringify(examination))
+            .map(response => response)
             .catch(this.handleErrorService.handleError);
     }
 
     findExaminationsBySubjectId(subjectId: number): Promise<SubjectExamination[]> {
-        return this.http.get(AppUtils.BACKEND_API_EXAMINATION_ALL_BY_SUBJECT_URL + '/' + subjectId)
+        return this.http.get<SubjectExamination[]>(AppUtils.BACKEND_API_EXAMINATION_ALL_BY_SUBJECT_URL + '/' + subjectId)
             .toPromise()
-            .then(response => response.json() as SubjectExamination[])
+            .then(response => response)
             .catch((error) => {
                 console.error('Error while getting examinations by subject id', error);
                 return Promise.reject(error.message || error);
