@@ -1,5 +1,5 @@
 import { Injectable } from "@angular/core";
-import { Http, Response } from "@angular/http";
+import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs/Observable';
 
 import { PatientsDicom, SerieDicom } from './dicom-data.model';
@@ -9,13 +9,11 @@ import { HandleErrorService } from '../shared/utils/handle-error.service';
 @Injectable()
 export class ImportService {
 
-    constructor(private http: Http, private handleErrorService: HandleErrorService) {
-        this.http = http;
-    }
+    constructor(private http: HttpClient, private handleErrorService: HandleErrorService) { }
 
     uploadFile(formData: FormData): Observable<PatientsDicom> {
-        return this.http.post(AppUtils.BACKEND_API_UPLOAD_DICOM_URL, formData)
-            .map(response => response.json() as PatientsDicom)
+        return this.http.post<PatientsDicom>(AppUtils.BACKEND_API_UPLOAD_DICOM_URL, formData)
+            .map(response => response)
             .catch(this.handleErrorService.handleError);
     }
 
@@ -25,6 +23,6 @@ export class ImportService {
             .catch((error) => {
                 console.error('Error while sending select series in json fromat', error);
                 return Promise.reject(error.message || error);
-        });
+            });
     }
 }  
